Move ChangeModel to Core
