+++ SyncKusto.Core/Models/FunctionSchemaDifference.cs
@@ -1,10 +1,9 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
 using SyncKusto.Core.Abstractions;
-using SyncKusto.Core.Models;
 
-namespace SyncKusto.ChangeModel
+namespace SyncKusto.Core.Models
 {
     public class FunctionSchemaDifference : SchemaDifference
     {
