+++ SyncKusto.Tests/DictionaryExtensionTests.cs
@@ -3,7 +3,7 @@
 
 using Bogus;
 using NUnit.Framework;
-using SyncKusto.Extensions;
+using SyncKusto.Core.Extensions;
 using System.Collections.Generic;
 using System.Linq;
